+++ PopupTabLayout/PopupTabLayout.js
@@ -9,7 +9,7 @@
  * @exports TabPanel
  */
 
-import {forKey, forProp, forward, forwardCustom, handle, stop} from '@enact/core/handle';
+import {forKey, forProp, forward, forwardCustom, handle, preventDefault, stop} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import useHandlers from '@enact/core/useHandlers';
 import {cap} from '@enact/core/util';
@@ -386,6 +386,11 @@ const tabPanelsHandlers = {
 			return document.querySelector(`section.${componentCss.body}`).contains(ev.target);
 		},
 		forwardCustom('onBack'),
+		() => {
+			Spotlight.setPointerMode(false);
+			return true;
+		},
+		preventDefault,
 		stop
 	)
 };