+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [2.5.1] - 2022-08-03
+
+### Added
+
+- `sandstone/Icon` supported icon list, adding a new icon `r2rappcall` 
+
+### Fixed
+
+- `sandstone/FixedPopupPanels` and `sandstone/PopupTabLayout` to restore scroll position when going back to the previous panel by left key
+- `sandstone/Panels.Panel` to restore focus properly when it has `sandstone/Scroller` with `focusableScrollbar`
+
 ## [2.5.0] - 2022-07-19
 
 ### Fixed